# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoLoop",
        "description": "Application de dons et d'échanges d'objets durables",
        "version": "1.0.0"
    },
    "paths": {
        "/api/annonces": {
            "get": {
                "tags": [
                    "Annonce"
                ],
                "summary": "liste de toutes les annonces disponibles",
                "operationId": "6d5efe14464468963c10a0a28ab8b88d",
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Ajout d'une annonce",
                "operationId": "24602bcd2a28bfa595b08a7a4acb2953",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "etat": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "categorie_id": {
                                        "type": "integer"
                                    },
                                    "localite_id": {
                                        "type": "integer"
                                    },
                                    "date_limite": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "image[]": {
                                        "description": "Liste de fichiers images",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "annonce ajoutée avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/annonces/admin": {
            "get": {
                "tags": [
                    "Annonce"
                ],
                "summary": "liste de toutes les annonces disponibles et non disponibles",
                "operationId": "4c623a557cdb0c578145cbc5557927ab",
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/annonces/{user}": {
            "get": {
                "tags": [
                    "Annonce"
                ],
                "summary": "les annonces d'un utilisateur",
                "operationId": "ddecb3cbe424586841a20cc20232db62",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/nombresAnnonces": {
            "get": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Nombres d'annonces",
                "operationId": "f0dc7afc7553cf8b0a4aa0aa7e8121c9",
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/annonces/{annonce}": {
            "get": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Details d'une annonce",
                "operationId": "20f3fa1ac5ce9b2456d092fe3660680c",
                "parameters": [
                    {
                        "name": "annonce",
                        "in": "path",
                        "description": "ID de l'annonce à afficher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            },
            "post": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Modification d'une annonce",
                "operationId": "3d6f4fdda2eec5462ac4cdc0fa34c902",
                "parameters": [
                    {
                        "name": "annonce",
                        "in": "path",
                        "description": "ID de l'annonce à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "etat": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "categorie_id": {
                                        "type": "integer"
                                    },
                                    "localite_id": {
                                        "type": "integer"
                                    },
                                    "date_limite": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "annonce modifiée avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cloturer/{annonce}": {
            "put": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Cloturer une annonce",
                "operationId": "5241a7be3a7ef9be03608f8be4922292",
                "parameters": [
                    {
                        "name": "annonce",
                        "in": "path",
                        "description": "ID de l'annonce à cloturer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/annoncesParCategorie/{categorie}": {
            "get": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Annonces par categorie",
                "operationId": "0c1fef97b3f1eef24b705fca7748b29e",
                "parameters": [
                    {
                        "name": "categorie",
                        "in": "path",
                        "description": "ID de la categorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/annoncesParLocalite/{localite}": {
            "get": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Annonces par localite",
                "operationId": "6af134c2844f80bd2ce6085b2adf4353",
                "parameters": [
                    {
                        "name": "localite",
                        "in": "path",
                        "description": "ID de la localite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/annoncesParType/{type}": {
            "get": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Annonces par Type",
                "operationId": "777525cce7b256f5868efa59790cc47e",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "type d'annonce entre Don ou Echange",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/annonce/{annonce}": {
            "delete": {
                "tags": [
                    "Annonce"
                ],
                "summary": "Supprimer une annonce",
                "operationId": "a6539c7ef620b16f39e852c256121535",
                "parameters": [
                    {
                        "name": "annonce",
                        "in": "path",
                        "description": "ID de l'annonce à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}